requirements.txt
google-generativeai
pandas
networkx
matplotlib